# Optimize VariableLookup#evaluate, Context#find_variable, Context#evaluate