--- a/ext/liquid_c/context.h
+++ b/ext/liquid_c/context.h
@@ -1,0 +1,7 @@
+#if !defined(LIQUID_CONTEXT_H)
+#define LIQUID_CONTEXT_H
+
+void init_liquid_context();
+
+#endif
+