--- a/ext/liquid_c/liquid.c
+++ b/ext/liquid_c/liquid.c
@@ -4,6 +4,7 @@
 #include "lexer.h"
 #include "parser.h"
 #include "block.h"
+#include "context.h"
 
 VALUE mLiquid, mLiquidC, cLiquidSyntaxError, cLiquidVariable, cLiquidTemplate;
 rb_encoding *utf8_encoding;
@@ -21,5 +22,6 @@ void Init_liquid_c(void)
     init_liquid_parser();
     init_liquid_variable();
     init_liquid_block();
+    init_liquid_context();
 }