--- a/ext/liquid_c/liquid.c
+++ b/ext/liquid_c/liquid.c
@@ -5,6 +5,7 @@
 #include "parser.h"
 #include "block.h"
 #include "context.h"
+#include "variable_lookup.h"
 
 VALUE mLiquid, mLiquidC, cLiquidSyntaxError, cLiquidVariable, cLiquidTemplate;
 rb_encoding *utf8_encoding;
@@ -23,5 +24,6 @@ void Init_liquid_c(void)
     init_liquid_variable();
     init_liquid_block();
     init_liquid_context();
+    init_liquid_variable_lookup();
 }