--- a/test/liquid_test.rb
+++ b/test/liquid_test.rb
@@ -5,7 +5,7 @@ $LOAD_PATH << liquid_test_dir
 require 'test_helper'
 require 'liquid/c'
 
-test_files = FileList[File.join(liquid_test_dir, 'integration/**/*_test.rb')]
+test_files = FileList[File.join(liquid_test_dir, '**/*_test.rb')]
 test_files.each do |test_file|
   require test_file
 end