This optimizes a hot path in rendering: the interaction of VariableLookup#evaluate, Context#find_variable, and Context#evaluate. With this change, we stay entirely within C for the whole blob of logic required to look up and evaluate variables. This produces the following change in render performance:

```
/opt/rubies/2.6.2/bin/ruby ./performance.rb bare benchmark strict

Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
              parse:     2.000  i/100ms
             render:     7.000  i/100ms
     parse & render:     1.000  i/100ms
Calculating -------------------------------------
              parse:     26.251  (± 3.8%) i/s -    262.000  in  10.001189s
             render:     77.849  (± 5.1%) i/s -    777.000  in  10.002960s
     parse & render:     18.779  (± 5.3%) i/s -    188.000  in  10.034829s

/opt/rubies/2.6.2/bin/ruby ./performance.rb c benchmark strict

Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
              parse:    10.000  i/100ms
             render:    10.000  i/100ms
     parse & render:     4.000  i/100ms
Calculating -------------------------------------
              parse:    104.356  (± 3.8%) i/s -      1.050k in  10.078616s
             render:    102.800  (± 4.9%) i/s -      1.030k in  10.046901s
     parse & render:     47.821  (± 4.2%) i/s -    480.000  in  10.058512s
```

This is the first set of render-related optimizations, so the `render` row can be compared directly. On a large Shopify merchant's theme, we see >20% reduction in Liquid rendering cost. 

@Shopify/liquid thoughts?