@pushrax 2019-04-11
> how does this relate to the changes that @fw42 suggested?

This change means we'll see less of a benefit of reducing the number of scopes (since iterating them is faster in C), but the changes should be complementary. We'll need to benchmark to find out.