@tobi 2019-04-11
So lets get this shipped, the idea of this other patch can speed up the c approach as well through another patch.