@tobi 2019-04-11
Let’s just fix that code in Shopify. That’s really bad idea

On Thu, Apr 11, 2019 at 5:22 PM Justin Li <notifications@github.com> wrote:

> *@pushrax* commented on this pull request.
> ------------------------------
>
> In ext/liquid_c/context.h
> <https://github.com/Shopify/liquid-c/pull/44#discussion_r274671060>:
>
> > +    if (klass == rb_cString || klass == rb_cArray || klass == rb_cHash)
>
> +        return value;
>
> +
>
> +    value = rb_funcall(value, id_to_liquid, 0);
>
> +
>
> +    if (rb_respond_to(value, id_set_context))
>
> +        rb_funcall(value, id_set_context, 1, context_to_set);
>
> +
>
> +    return value;
>
> +}
>
> +
>
> +
>
> +inline static VALUE materialize_proc(VALUE context, VALUE scope, VALUE key, VALUE value)
>
> +{
>
> +    if (scope != Qnil && !RB_SPECIAL_CONST_P(value) && TYPE(scope) == T_HASH) {
>
> +        if (rb_obj_is_kind_of(value, rb_cProc)) {
>
>
> Possibly a breaking change – I know some code in Shopify creates objects
> that override #is_a?(Proc) to return true. Calling into Ruby here just to
> keep that behaviour is really not ideal though.
>
> —
> You are receiving this because you are on a team that was mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Shopify/liquid-c/pull/44#pullrequestreview-225800237>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAABW3MGMSH3Sip4Kd7Rw12r02jv7U1oks5vf6eXgaJpZM4cqsGl>
> .
>
-- 
- tobi
CEO Shopify