@pushrax 2019-04-23
All comments have been addressed, and this is passing in Shopify core, bump for final review.